Upgrade neva-main-rolling from 151.0.7913.0~1 to 151.0.7922.0~1.

Conflicts:
	build/config/c++/c++.gni

Issue: NEVA-11319

+++ chrome/browser/ui/views/translate/translate_bubble_view_browsertest.cc
@@ -317,7 +317,8 @@ IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIDisabled,
                                                            u"Spanish");
 
   EXPECT_EQ(bubble->model()->GetTargetLanguageIndex(),
-            bubble->model()->GetTargetLanguageIndexForCode("es").value());
+            static_cast<int>(
+                bubble->model()->GetTargetLanguageIndexForCode("es").value()));
 }
 
 IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIEnabled,
@@ -338,7 +339,8 @@ IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIEnabled,
                                                            u"Spanish");
 
   EXPECT_EQ(bubble->model()->GetTargetLanguageIndex(),
-            bubble->model()->GetTargetLanguageIndexForCode("es").value());
+            static_cast<int>(
+                bubble->model()->GetTargetLanguageIndexForCode("es").value()));
 }
 
 IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIEnabled,
